# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order Notifier API",
        "description": "API for managing orders with notifications",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.example.com/api/v1",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List orders with filters",
                "description": "Get all orders for authenticated user with dynamic filters using Pipeline pattern. Requires orders:read ability.",
                "operationId": "5894f7ae19272fefaf73e3a6f9d9782d",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (comma-separated)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "placed,processing"
                        }
                    },
                    {
                        "name": "min_amount",
                        "in": "query",
                        "description": "Minimum amount filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 100
                        }
                    },
                    {
                        "name": "max_amount",
                        "in": "query",
                        "description": "Maximum amount filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 500
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search in order code",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "ORD-1001"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter orders from date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-10-01"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter orders until date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-10-31"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort by field (prefix with - for desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "-placed_at"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "orders": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "ORD-1001"
                                                    },
                                                    "amount_decimal": {
                                                        "type": "number",
                                                        "example": 199.99
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "placed"
                                                    },
                                                    "placed_at": {
                                                        "type": "string",
                                                        "format": "datetime"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - missing orders:read ability"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create new order",
                "description": "Create a new order for authenticated user. Requires orders:write ability.",
                "operationId": "fccc497d8bcac727afdc854cf0521698",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "amount_decimal",
                                    "status",
                                    "placed_at"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "ORD-1001"
                                    },
                                    "amount_decimal": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 199.99
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "placed",
                                            "processing",
                                            "shipped",
                                            "delivered",
                                            "cancelled"
                                        ],
                                        "example": "placed"
                                    },
                                    "placed_at": {
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2025-10-29 10:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "order": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - missing orders:write ability"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order details",
                "description": "Get single order by ID. User can only access their own orders. Requires orders:read ability.",
                "operationId": "37b9054f8bb7c660ae6ee13bfb030a08",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "order": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - not your order or missing orders:read ability"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update order status",
                "description": "Update order status. Triggers FCM push notification on status change. Requires orders:write ability.",
                "operationId": "9e1199abc9156a9e0890092130c43b48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "placed",
                                            "processing",
                                            "shipped",
                                            "delivered",
                                            "cancelled"
                                        ],
                                        "example": "shipped"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated and notification queued",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "order": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - not your order or missing orders:write ability"
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}